+++ src/save_data/mass_effect_1_le/player.rs
@@ -34,16 +34,33 @@ pub struct Player {
     pub grenades: f32,
     pub omnigel: f32,
     pub face_code: ImguiString,
-    _unknown4: Dummy<4>,
+    armor_overridden: bool,
     auto_levelup_template_id: i32,
     health_per_level: f32,
-    _unknown5: Dummy<9>,
+    stability_current: f32,
+    race: u8,
+    toxic_current: f32,
     stamina: i32,
     focus: i32,
     precision: i32,
     coordination: i32,
-    _unknown6: Dummy<14>,
+    attribute_primary: u8,
+    attribute_secondary: u8,
+    skill_charm: f32,
+    skill_intimidate: f32,
+    skill_haggle: f32,
     health_current: f32,
+    shield_current: f32,
+    xp_level: i32,
+    is_driving: bool,
+    pub game_options: Vec<i32>,
+    helmet_shown: bool,
+    _unknown: Dummy<5>,
+    last_power: ImguiString,
+    health_max: f32,
+    hotkeys: Vec<Hotkey>,
+    primary_weapon: ImguiString,
+    secondary_weapon: ImguiString,
 }
 
 #[derive(Deserialize, Serialize, RawUi, Clone, Default)]
@@ -69,7 +86,7 @@ pub struct Inventory {
     pub equipment: Vec<Item>,
     pub quick_slots: Vec<Item>,
     pub inventory: Vec<Item>,
-    plot_items: Vec<Item>,
+    buy_pack_items: Vec<Item>, // TODO: Inventory tab
 }
 
 #[allow(clippy::upper_case_acronyms)]
@@ -111,4 +128,10 @@ struct ItemMod {
     item_level: ItemLevel,
     manufacturer_id: i32,
     plot_conditional_id: i32,
+}
+
+#[derive(Deserialize, Serialize, RawUi, Clone, Default)]
+struct Hotkey {
+    pawn: i32,
+    event: i32,
 }